+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-- [ANF-03](https://osklenbr.atlassian.net/browse/ANF-03) - Configure formatting, Prisma, and project docs
+- [ANF-03](https://osklenbr.atlassian.net/browse/ANF-03) - Scaffold NestJS API, Prisma, Docker, Auth.
 - [ANF-02](https://osklenbr.atlassian.net/browse/ANF-02) - Organize domain storytelling suggestions.
 - [ANF-01] - Scaffold NestJS starter project